VT-635 oops

+++ valintaperusteet-service/src/main/resources/db/migration/V014__add_indeksi_column_to_valintaperuste_viite_and_constraints.sql
@@ -1,4 +1,4 @@
-    alter table valintaperuste_viite add column indeksi int8;
+    alter table valintaperuste_viite add column indeksi int4;
     update valintaperuste_viite set indeksi = 1;
 
     alter table valintaperuste_viite alter column indeksi set not null;
